+++ core/src/net/oschina/j2cache/ClusterPolicyFactory.java
@@ -15,7 +15,6 @@
  */
 package net.oschina.j2cache;
 
-import net.oschina.j2cache.redis.RedisClient;
 import net.oschina.j2cache.redis.RedisPubSubClusterPolicy;
 
 import java.util.Properties;
@@ -29,12 +28,11 @@ public class ClusterPolicyFactory {
     /**
      * 使用 Redis 订阅和发布机制，该方法只能调用一次
      * @param name  频道名称
-     * @param redis Redis 客户端接口
      * @param props 框架配置
      * @return 返回 Redis 集群策略的实例
      */
-    public final static ClusterPolicy redis(String name, RedisClient redis, Properties props) {
-        RedisPubSubClusterPolicy policy = new RedisPubSubClusterPolicy(name, redis);
+    public final static ClusterPolicy redis(String name, Properties props) {
+        RedisPubSubClusterPolicy policy = new RedisPubSubClusterPolicy(name, props);
         policy.connect(props);
         return policy;
     }